// src/database/dbInterfaces/database.interfaces.ts
export enum ISQLite3TableNames {
  USERS_TABLE = 'users',
  PRODUCTS_TABLE = 'products',
  MATERIALS = 'materials',
  RECIPES_TABLE = 'recipes',
  RECIPE__ITEMS_TABLE = 'recipe_items',
  SHELF_STOCK_TRANSACTIONS_TABLE = 'shelf_stock_transactions',
  STOCK_TRANSACTIONS_TABLE = 'stock_transactions'
}